requirements.txt
backcall==0.1.0
cx-Freeze==5.1.1
decorator==4.3.0
ipython-genutils==0.2.0
jedi==0.12.0
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyQt5==5.10.1
simplegeneric==0.8.1
sip==4.19.8
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
